requirements.txt
pandas
requests
bs4
google-api-python-client
pdfplumber
